add `EndpointResult::IncomingError`
Allows the user to respond to the `Incoming::accept()` process failing.

`Incoming::accept()` failures are likely not caused by the application or the remote. The QUIC connection listens on a normal UDP socket and any reachable network endpoint can send datagrams to it, solicited or not. The data received might not even be a QUIC packet.

This should not be a fatal error, and users should just log and move on.